Update python-hjson to 3.1.0
Unfortunately, the GitHub repository for python-hjson does not contain a
proper tag for the v3.1.0 release; however, building the package from
the PyPi sources does not work, because those sources are not complete
(they don't contain data files for tests, and Guix usually runs tests
during the build).  As a workaround, hardcode the hash of the latest
commit, which has the message of "Publish 3.1.0"; the sources in that
commit seems to match the contents of the tarball from PyPi.