Add a CI workflow to check that the manifest works
Sometimes Guix developers make changes which break compatibility; try
using the GitHub CI to catch those cases when they happen.

Unfortunately, getting Guix to work in a GitHub workflow is really
complicated, and a straightforward way to do that is not practical
because of major performance problems, therefore this CI workflow saves
the Guix store contents in the GitHub cache and runs `guix pull` only
once a day (subsequent runs on the same day reuse the cached store).

The first CI run (when the cache is initially empty) may need multiple
attempts to succeed, because sometimes the `guix substitute` processes
fail after a long build; the Guix store state will be saved even after
such failures, so that the next attempt could continue from that state
and eventually succeed.